# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://notin.ir/api",
            "description": "Production Server"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/wallet": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "دریافت موجودی کیف پول",
                "description": "بازیابی اطلاعات کیف پول و موجودی کوین‌های کاربر فعلی",
                "operationId": "32f3d42f4d08f2f58f709012d7d29d63",
                "responses": {
                    "200": {
                        "description": "اطلاعات کیف پول دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "coins": {
                                                    "description": "موجودی کوین در کیف پول",
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "کیف پول پیدا نشد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "کیف پول یافت نشد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "",
                "bearerFormat": "",
                "scheme": "bearer"
            }
        }
    },
...
}